add middleware, update render(), dash template, alpine menus
