+++ app/static/main.css
@@ -485,148 +485,6 @@ a {
   color: #000;
 }
 
-/* header */
-
-.dashboard-header {
-  z-index: 3;
-}
-
-.dashboard-header ul {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  overflow: hidden;
-}
-
-.dashboard-header li a {
-  padding: 0.5rem;
-  text-decoration: none;
-}
-
-.dashboard-header li a:hover,
-.dashboard-header .menu-btn:hover {}
-
-.dashboard-header .logo {
-  display: block;
-  float: left;
-  font-size: 1.2rem;
-  padding: 0.75rem 0.25rem;
-  text-decoration: none;
-}
-
-/* menu */
-
-.dashboard-header .menu {
-  clear: both;
-  max-height: 0;
-  width: 100%;
-  transition: max-height .2s ease-out;
-}
-
-.dashboard-header .menu li {
-  margin: 0.5rem 0;
-}
-
-/* menu icon */
-
-.dashboard-header .menu-icon {
-  cursor: pointer;
-  display: inline-block;
-  float: right;
-  padding: 24px 20px;
-  position: relative;
-  user-select: none;
-}
-
-.dashboard-header .menu-icon .navicon {
-  background: #333;
-  display: block;
-  height: 2px;
-  position: relative;
-  transition: background .2s ease-out;
-  width: 18px;
-}
-
-.dashboard-header .menu-icon .navicon:before,
-.dashboard-header .menu-icon .navicon:after {
-  background: #333;
-  content: '';
-  display: block;
-  height: 100%;
-  position: absolute;
-  transition: all .2s ease-out;
-  width: 100%;
-}
-
-.dashboard-header .menu-icon .navicon:before {
-  top: 5px;
-}
-
-.dashboard-header .menu-icon .navicon:after {
-  top: -5px;
-}
-
-/* menu btn */
-
-.dashboard-header .menu-btn {
-  margin: 1rem;
-  display: none;
-}
-
-.dashboard-header .menu {
-  margin: 1rem;
-  display: none;
-}
-
-.dashboard-header .menu-btn:checked~.menu {
-  margin: 1rem;
-  max-height: 600px;
-  display: block;
-}
-
-.dashboard-header .menu-btn:checked~.menu-icon .navicon {
-  background: transparent;
-}
-
-.dashboard-header .menu-btn:checked~.menu-icon .navicon:before {
-  transform: rotate(-45deg);
-}
-
-.dashboard-header .menu-btn:checked~.menu-icon .navicon:after {
-  transform: rotate(45deg);
-}
-
-.dashboard-header .menu-btn:checked~.menu-icon:not(.steps) .navicon:before,
-.dashboard-header .menu-btn:checked~.menu-icon:not(.steps) .navicon:after {
-  top: 0;
-}
-
-
-@media (min-width: 640px) {
-  .dashboard-header li {
-    float: left;
-  }
-
-
-  .dashboard-header .menu {
-    padding-left: 1rem;
-    clear: none;
-    float: right;
-    max-height: none;
-    margin: 1rem;
-    display: block;
-  }
-
-  .dashboard-header .menu-icon {
-    display: none;
-  }
-}
-
-
-
-
-
-
 
 @media(prefers-reduced-motion) {
   .hide,
