+++ app/views/components/footer.gohtml
@@ -1,8 +1,4 @@
 {{define "footer"}}
-
-</div>
-</div>
-<!-- Footer -->
 <footer class="pt-24 pb-12 flex grow place-items-end bottom-0 w-full text-gray-900 px-8 sm:px-14">
 	<div class="w-full max-w-4xl mx-auto">
 		<div class="w-full span-2">
@@ -70,7 +66,4 @@
 	</div>
 </div>
 </footer>
-</body>
-
-</html>
 {{end}}
