+++ app/views/components/head.gohtml
@@ -1,0 +1,49 @@
+{{define "head"}}
+<head>
+
+	<!-- Primary -->
+	<title>{{if .Title}}{{.Title}}{{else}}Kevin Gil{{end}}</title>
+	<meta name="title" content="{{if .Title}}{{.Title}}{{else}}Kevin Gil{{end}}" />
+	<meta name="description" content="Software Engineer in San Francisco"/>
+
+	<!-- Open Graph -->
+	<meta property="og:type" content="Blog" />
+	<meta property="og:url" content="kevingil.com" />
+	<meta property="og:title" content="{{if .Title}}{{.Title}}{{else}}Kevin Gil{{end}}" />
+	<meta property="og:description" content="Software Engineer in San Francisco" />
+	<meta property="og:image" content="/static/og-image.png" />
+	
+	<!-- CSS/JS -->
+	<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
+	<script src="/static/tailwind-3-33.css"></script>
+	<script src="/static/htmx.min.js"></script>
+	<script src="/static/main.js?v={{v}}"></script>
+	<script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js"></script>
+	<script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
+
+	<!-- Favicon -->
+	<link rel="apple-touch-icon" sizes="180x180" href="/static/apple-touch-icon.png">
+	<link rel="icon" type="image/png" sizes="32x32" href="/static/favicon-32x32.png">
+	<link rel="icon" type="image/png" sizes="16x16" href="/static/favicon-16x16.png">
+	<link rel="stylesheet" href="/static/prism/prism.css">
+	<link rel="stylesheet" href="/static/main.css?v={{v}}">
+	<!-- Font Awesome -->
+	<script src="https://kit.fontawesome.com/99fbdcde4b.js" crossorigin="anonymous"></script>
+	<!-- Fonts -->
+	<link rel="preconnect" href="https://fonts.googleapis.com">
+	<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+	<link
+		href="https://fonts.googleapis.com/css2?family=Inter:wght@100;300;400;500;700&family=Nanum+Gothic+Coding:wght@400;700&display=swap"
+		rel="stylesheet">
+
+</head>
+
+<script async src="https://www.googletagmanager.com/gtag/js?id=G-RV6Q7F8Z8K"></script>
+<script>
+	window.dataLayer = window.dataLayer || [];
+	function gtag() { dataLayer.push(arguments); }
+	gtag('js', new Date());
+
+	gtag('config', 'G-RV6Q7F8Z8K');
+</script>
+{{end}}
