--- app/views/components/header.gohtml
@@ -1,89 +1,0 @@
-{{define "header"}}
-<!DOCTYPE html>
-<html lang="en" class="light" style="color-scheme: light;">
-
-<head>
-
-	<!-- Primary -->
-	<title>{{if .Title}}{{.Title}}{{else}}Kevin Gil{{end}}</title>
-	<meta name="title" content="{{if .Title}}{{.Title}}{{else}}Kevin Gil{{end}}" />
-	<meta name="description" content="Software Engineer in San Francisco"/>
-
-	<!-- Open Graph -->
-	<meta property="og:type" content="Blog" />
-	<meta property="og:url" content="kevingil.com" />
-	<meta property="og:title" content="{{if .Title}}{{.Title}}{{else}}Kevin Gil{{end}}" />
-	<meta property="og:description" content="Software Engineer in San Francisco" />
-	<meta property="og:image" content="/static/og-image.png" />
-	
-	<!-- CSS/JS -->
-	<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">
-	<script src="/static/tailwind-3-33.css"></script>
-	<script src="/static/htmx.min.js"></script>
-	<script src="/static/main.js?v={{v}}"></script>
-	<script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js"></script>
-	<script src="https://cdn.jsdelivr.net/npm/marked/marked.min.js"></script>
-
-	<!-- Favicon -->
-	<link rel="apple-touch-icon" sizes="180x180" href="/static/apple-touch-icon.png">
-	<link rel="icon" type="image/png" sizes="32x32" href="/static/favicon-32x32.png">
-	<link rel="icon" type="image/png" sizes="16x16" href="/static/favicon-16x16.png">
-	<link rel="stylesheet" href="/static/prism/prism.css">
-	<link rel="stylesheet" href="/static/main.css?v={{v}}">
-	<!-- Font Awesome -->
-	<script src="https://kit.fontawesome.com/99fbdcde4b.js" crossorigin="anonymous"></script>
-	<!-- Fonts -->
-	<link rel="preconnect" href="https://fonts.googleapis.com">
-	<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-	<link
-		href="https://fonts.googleapis.com/css2?family=Inter:wght@100;300;400;500;700&family=Nanum+Gothic+Coding:wght@400;700&display=swap"
-		rel="stylesheet">
-
-</head>
-<script async src="https://www.googletagmanager.com/gtag/js?id=G-RV6Q7F8Z8K"></script>
-<script>
-	window.dataLayer = window.dataLayer || [];
-	function gtag() { dataLayer.push(arguments); }
-	gtag('js', new Date());
-
-	gtag('config', 'G-RV6Q7F8Z8K');
-</script>
-
-<body class=" flex min-h-screen flex flex-col items-center">
-	<nav x-data="{ isOpen: false, isLargeScreen: window.innerWidth >= 640 }" 
-		x-init="() => { window.addEventListener('resize', () => { isLargeScreen = window.innerWidth >= 640 }) }" 
-		:class="{ 'bg-white': isOpen && !isLargeScreen }"
-		class="fixed top-0 z-10 scrollfade backdrop-blur-sm w-full">
-		<div class="container max-w-4xl py-4 px-3 mx-auto flex flex-wrap items-center justify-between">
-		  <a id="logosvg" hx-get="/" hx-push-url="true" hx-boost="true" hx-swap="innerHTML transition:true show:window:top" hx-target="#container" class="flex items-center text-2xl font-semibold">
-			Kevin Gil
-		  </a>
-		  <div @click="isOpen = true" class="sm:hidden">
-		  <button  class="flex items-center px-3 py-2 border rounded-md text-gray-900 border-gray-900 hover:text-indigo-600 hover:border-indigo-600">
-			<svg class="fill-current h-3 w-3" viewBox="0 0 20 20" xmlns="http://www.w3.org/2000/svg">
-			  <title>Menu</title>
-			  <path x-show="!isOpen" fill-rule="evenodd" d="M3 5a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zM3 10a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zM3 15a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1z" clip-rule="evenodd" />
-			  <path x-show="isOpen" fill-rule="evenodd" d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z" clip-rule="evenodd" />
-			</svg>
-		  </button>
-		  </div>
-		  <div x-show="isOpen || isLargeScreen" x-cloak @click.away="isOpen = false"
-		    class="w-full sm:w-auto sm:inline-block sm:ml-auto mt-4 sm:mt-0 sm:flex sm:flex-row sm:items-center">
-			<div class="flex flex-col sm:flex-row gap-6 font-medium rounded-lg sm:border-0 list-none">
-			  <div>
-				<a @click="isOpen = false" hx-get="/blog" hx-boost="true" hx-swap="innerHTML transition:true show:window:top" hx-target="#container" hx-push-url="true" class="block rounded sm:border-0 font-semibold hover:text-indigo-600 p-0 text-gray-900">Blog</a>
-			  </div>
-			  <div>
-				<a @click="isOpen = false" hx-get="/about" hx-boost="true" hx-swap="innerHTML transition:true show:window:top" hx-target="#container" hx-push-url="true" class="block rounded sm:border-0 font-semibold hover:text-indigo-600 p-0 text-gray-900">About</a>
-			  </div>
-			  <div>
-				<a @click="isOpen = false" hx-get="/contact" hx-boost="true" hx-swap="innerHTML transition:true show:window:top" hx-target="#container" hx-push-url="true" class="block rounded sm:border-0 font-semibold hover:text-indigo-600 p-0 text-gray-900">Contact</a>
-			  </div>
-			</div>
-		  </div>
-		</div>
-	  </nav>
-
-	<div id="main" class="w-full pt-[75px]">
-		<div id="container" class="slide-it p-0">
-			{{end}}
