+++ app/views/index.gohtml
@@ -66,7 +66,7 @@
     </section>
 
     <section class="px-2 pb-16 ">
-        <div class="flex flex-col sm:flex-row gap-4 max-w-4xl mx-auto">
+        <div class="flex flex-col sm:flex-row gap-4 mx-auto">
             <div class="fade-c1 shad p-4 w-full rounded-lg">
                 <i class="fa-brands fa-github text-2xl"></i>
                 <h3 class="font-bold">Github</h3>
