+++ app/views/layout.gohtml
@@ -1,7 +1,19 @@
 {{define "layout"}}
-    {{template "header"}}	
-    {{ if .}}
-        {{ .}}
-    {{end}}
-    {{template "footer"}}
+<!DOCTYPE html>
+<html lang="en" class="light" style="color-scheme: light;">
+    {{template "head"}}
+    <body class=" flex min-h-screen flex flex-col items-center">
+        {{ template "navigation" .}}
+        <div id="main" class="w-full pt-[75px]">
+            <div id="container" class="slide-it p-0 mx-auto max-w-4xl">
+                {{ if .TemplateChild}}
+                    {{ .TemplateChild}}
+                {{else}}
+                <p class="p-4 bg-white">Error rendering child.</p>
+                {{end}}
+            </div>
+        </div>
+        {{template "footer" .}}
+    </body>
+</html>
 {{end}}
