+++ app/views/pages/about.gohtml
@@ -1,10 +1,10 @@
 {{define "about"}}
 <!-- About Section -->
 <section class="w-full">
-	<article class="flex flex-col text-zinc-700 mx-auto max-w-4xl">
+	<article class="flex flex-col text-zinc-700 mx-auto">
 		<div class="flex flex-col justify-start">
 			<p class="text-5xl text-indigo-500 py-6 pb-10 px-4 sm:px-6 hide">About.</p>
-			<div class="mx-auto w-full max-w-4xl p-4 sm:p-6 rounded-xl bg-white shad">
+			<div class="mx-auto w-full p-4 sm:p-6 rounded-xl bg-white shad">
 				<div class="">
 					<img src="/static/images/profile.jpg" class="hide-card rounded-full mx-auto max-h-28 sm:max-h-40 mt-4 mb-16 delay-500 shad" />
 				</div>
