+++ app/views/pages/blog.gohtml
@@ -1,6 +1,6 @@
 {{define "blog"}}
 <!-- Articles Section -->
-<section class="w-full w-full max-w-4xl mx-auto">
+<section class="w-full w-full mx-auto">
     <p class="text-5xl text-indigo-500 py-6 pb-10 px-6 hide">Blog.</p>
     {{if not .Articles}}
     <p>No articles.</p>
