+++ app/views/pages/contact.gohtml
@@ -1,7 +1,7 @@
 {{define "contact"}}
 <!-- Contact Section -->
 <section  class="w-full">
-	<article class="flex flex-col text-zinc-800 mx-auto max-w-4xl">
+	<article class="flex flex-col text-zinc-800 mx-auto">
 		<div class="flex flex-col justify-start">
 			<p class="text-5xl text-indigo-500 py-6 pb-10 px-4 sm:px-6 hide">Contact.</p>
 			<div class="mx-auto w-full p-4 sm:p-6 rounded-xl bg-white shad">
