+++ app/views/pages/dashboard-layout.gohtml
@@ -1,85 +1,25 @@
 {{define "dashboard-layout"}}
-{{template "header"}}
-<div class="grid grid-cols-1 gap-0 sm:gap-2 sm:grid-cols-5 max-w-4xl mx-auto">
-	<div class="dashboard-header">
-		<p class="logo">Dashboard</p>
-		<input class="menu-btn" type="checkbox" id="menu-btn" />
-		<label class="menu-icon" for="menu-btn"><span class="navicon"></span></label>
-		<div class="menu">
-			<div class="w-full">
-				<a hx-get="/dashboard" hx-boost="true" hx-swap="innerHTML" hx-target="#dash_container"
-				hx-push-url="true" class="flex gap-2 rounded-md py-2 pl-3 hover:bg-zinc-200">
-					<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
-						stroke="currentColor" class="w-6 h-6">
-						<path stroke-divnecap="round" stroke-divnejoin="round"
-							d="M2.25 12l8.954-8.955c.44-.439 1.152-.439 1.591 0L21.75 12M4.5 9.75v10.125c0 .621.504 1.125 1.125 1.125H9.75v-4.875c0-.621.504-1.125 1.125-1.125h2.25c.621 0 1.125.504 1.125 1.125V21h4.125c.621 0 1.125-.504 1.125-1.125V9.75M8.25 21h8.25" />
-					</svg>
-					Home
-				</a>
-			</div>
-			<div class="w-full">
-				<a hx-get="/dashboard/profile" hx-boost="true" hx-swap="innerHTML" hx-target="#dash_container"
-					hx-push-url="true" class="flex gap-2 rounded-md py-2 pl-3 hover:bg-zinc-200">
-					<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
-						stroke="currentColor" class="w-6 h-6">
-						<path stroke-divnecap="round" stroke-divnejoin="round"
-							d="M17.982 18.725A7.488 7.488 0 0012 15.75a7.488 7.488 0 00-5.982 2.975m11.963 0a9 9 0 10-11.963 0m11.963 0A8.966 8.966 0 0112 21a8.966 8.966 0 01-5.982-2.275M15 9.75a3 3 0 11-6 0 3 3 0 016 0z" />
-					</svg>
-					User</a>
-			</div>
-			<div class="w-full">
-				<a hx-get="/dashboard/publish" hx-boost="true" hx-swap="innerHTML" hx-target="#dash_container"
-					hx-push-url="true" class="flex gap-2 rounded-md py-2 pl-3 hover:bg-zinc-200">
-					<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-						<path stroke-divnecap="round" stroke-divnejoin="round" d="M12 7.5h1.5m-1.5 3h1.5m-7.5 3h7.5m-7.5 3h7.5m3-9h3.375c.621 0 1.125.504 1.125 1.125V18a2.25 2.25 0 0 1-2.25 2.25M16.5 7.5V18a2.25 2.25 0 0 0 2.25 2.25M16.5 7.5V4.875c0-.621-.504-1.125-1.125-1.125H4.125C3.504 3.75 3 4.254 3 4.875V18a2.25 2.25 0 0 0 2.25 2.25h13.5M6 7.5h3v3H6v-3Z" />
-					  </svg>
-					Articles
-				</a>
-			</div>
-			<div class="w-full">
-				<a hx-get="/dashboard/resume" hx-boost="true" hx-swap="innerHTML" hx-target="#dash_container"
-					hx-push-url="true" class="flex gap-2 rounded-md py-2 pl-3 hover:bg-zinc-200">
-					<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
-						stroke="currentColor" class="w-6 h-6">
-						<path stroke-divnecap="round" stroke-divnejoin="round"
-							d="M19.5 14.25v-2.625a3.375 3.375 0 00-3.375-3.375h-1.5A1.125 1.125 0 0113.5 7.125v-1.5a3.375 3.375 0 00-3.375-3.375H8.25m0 12.75h7.5m-7.5 3H12M10.5 2.25H5.625c-.621 0-1.125.504-1.125 1.125v17.25c0 .621.504 1.125 1.125 1.125h12.75c.621 0 1.125-.504 1.125-1.125V11.25a9 9 0 00-9-9z" />
-					</svg>
-					Resume</a>
-			</div>
-			<div class="w-full">
-				<a hx-get="/dashboard/files" hx-boost="true" hx-swap="innerHTML" hx-target="#dash_container"
-					hx-push-url="true" class="flex gap-2 rounded-md py-2 pl-3 hover:bg-zinc-200">
-					<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
-						stroke="currentColor" class="w-6 h-6">
-						<path stroke-divnecap="round" stroke-divnejoin="round"
-							d="M2.25 15.75l5.159-5.159a2.25 2.25 0 013.182 0l5.159 5.159m-1.5-1.5l1.409-1.409a2.25 2.25 0 013.182 0l2.909 2.909m-18 3.75h16.5a1.5 1.5 0 001.5-1.5V6a1.5 1.5 0 00-1.5-1.5H3.75A1.5 1.5 0 002.25 6v12a1.5 1.5 0 001.5 1.5zm10.5-11.25h.008v.008h-.008V8.25zm.375 0a.375.375 0 11-.75 0 .375.375 0 01.75 0z" />
-					</svg>
-					Files</a>
-			</div>
-			<div class="w-full">
-				<a href="/logout" class="flex gap-2 rounded-md py-2 pl-3 hover:bg-zinc-200">
-					<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
-						stroke="currentColor" class="w-6 h-6">
-						<path stroke-divnecap="round" stroke-divnejoin="round"
-							d="M15.75 9V5.25A2.25 2.25 0 0013.5 3h-6a2.25 2.25 0 00-2.25 2.25v13.5A2.25 2.25 0 007.5 21h6a2.25 2.25 0 002.25-2.25V15M12 9l-3 3m0 0l3 3m-3-3h12.75" />
-					</svg>
+<!DOCTYPE html>
+<html lang="en" class="light" style="color-scheme: light;">
+{{template "head"}}
 
-					Logout</a>
-			</div>
+<body class=" flex min-h-screen flex flex-col items-center">
+	{{ template "navigation" .}}
+	<div id="main" class="w-full pt-[75px]">
+		<div id="container" class="slide-it p-0 mx-auto max-w-4xl">
+				{{ if .TemplateChild}}
+
+
+				{{ .TemplateChild}}
+
+
+				{{else}}
+				<p class="p-4 bg-white">Error rendering child.</p>
+				{{end}}
 		</div>
 	</div>
-	<script>
-		document.querySelectorAll('.menu div a').forEach(function (a) {
-			a.addEventListener('click', function () {
-				document.getElementById('menu-btn').checked = false;
-			});
-		});
-	</script>
-		
-	</script>
-	<div id="dash_container" class="sm:col-span-4">
-		{{ .}}
-	</div>
-</div>
-{{template "footer"}}
+	{{template "footer"}}
+</body>
+
+</html>
 {{end}}
