+++ app/views/pages/login.gohtml
@@ -1,33 +1,44 @@
-{{template "header"}}
-		<div class="mt-16 p-4 max-w-[400px] mx-auto">
+{{define "login"}}
+<div class="mt-16 p-4 max-w-[400px] mx-auto">
+	<div>
+		<p class="text-left text-3xl font-semibold flex gap-2">
+			Publish
+			<svg xmlns="http://www.w3.org/2000/svg" class="w-5 pb-[0.25rem]" viewBox="0 0 512 512">
+				<path
+					d="M57.7 193l9.4 16.4c8.3 14.5 21.9 25.2 38 29.8L163 255.7c17.2 4.9 29 20.6 29 38.5v39.9c0 11 6.2 21 16 25.9s16 14.9 16 25.9v39c0 15.6 14.9 26.9 29.9 22.6c16.1-4.6 28.6-17.5 32.7-33.8l2.8-11.2c4.2-16.9 15.2-31.4 30.3-40l8.1-4.6c15-8.5 24.2-24.5 24.2-41.7v-8.3c0-12.7-5.1-24.9-14.1-33.9l-3.9-3.9c-9-9-21.2-14.1-33.9-14.1H257c-11.1 0-22.1-2.9-31.8-8.4l-34.5-19.7c-4.3-2.5-7.6-6.5-9.2-11.2c-3.2-9.6 1.1-20 10.2-24.5l5.9-3c6.6-3.3 14.3-3.9 21.3-1.5l23.2 7.7c8.2 2.7 17.2-.4 21.9-7.5c4.7-7 4.2-16.3-1.2-22.8l-13.6-16.3c-10-12-9.9-29.5 .3-41.3l15.7-18.3c8.8-10.3 10.2-25 3.5-36.7l-2.4-4.2c-3.5-.2-6.9-.3-10.4-.3C163.1 48 84.4 108.9 57.7 193zM464 256c0-36.8-9.6-71.4-26.4-101.5L412 164.8c-15.7 6.3-23.8 23.8-18.5 39.8l16.9 50.7c3.5 10.4 12 18.3 22.6 20.9l29.1 7.3c1.2-9 1.8-18.2 1.8-27.5zM0 256a256 256 0 1 1 512 0A256 256 0 1 1 0 256z" />
+			</svg>
+		</p>
+		<p class="text-left text-sm text-gray-500 hide">Hello 😉</p>
+	</div>
+	<form class="mt-8 space-y-6" action="/login" method="POST">
+		<div class="rounded-md shadow-sm -space-y-px">
 			<div>
-				<p class="text-left text-3xl font-semibold flex gap-2">
-					Publish
-					<svg xmlns="http://www.w3.org/2000/svg" class="w-5 pb-[0.25rem]" viewBox="0 0 512 512"><path d="M57.7 193l9.4 16.4c8.3 14.5 21.9 25.2 38 29.8L163 255.7c17.2 4.9 29 20.6 29 38.5v39.9c0 11 6.2 21 16 25.9s16 14.9 16 25.9v39c0 15.6 14.9 26.9 29.9 22.6c16.1-4.6 28.6-17.5 32.7-33.8l2.8-11.2c4.2-16.9 15.2-31.4 30.3-40l8.1-4.6c15-8.5 24.2-24.5 24.2-41.7v-8.3c0-12.7-5.1-24.9-14.1-33.9l-3.9-3.9c-9-9-21.2-14.1-33.9-14.1H257c-11.1 0-22.1-2.9-31.8-8.4l-34.5-19.7c-4.3-2.5-7.6-6.5-9.2-11.2c-3.2-9.6 1.1-20 10.2-24.5l5.9-3c6.6-3.3 14.3-3.9 21.3-1.5l23.2 7.7c8.2 2.7 17.2-.4 21.9-7.5c4.7-7 4.2-16.3-1.2-22.8l-13.6-16.3c-10-12-9.9-29.5 .3-41.3l15.7-18.3c8.8-10.3 10.2-25 3.5-36.7l-2.4-4.2c-3.5-.2-6.9-.3-10.4-.3C163.1 48 84.4 108.9 57.7 193zM464 256c0-36.8-9.6-71.4-26.4-101.5L412 164.8c-15.7 6.3-23.8 23.8-18.5 39.8l16.9 50.7c3.5 10.4 12 18.3 22.6 20.9l29.1 7.3c1.2-9 1.8-18.2 1.8-27.5zM0 256a256 256 0 1 1 512 0A256 256 0 1 1 0 256z"/></svg>
-				</p>
-				<p class="text-left text-sm text-gray-500 hide">Hello 😉</p>
+				<label for="email" class="sr-only">Email</label>
+				<input id="email" name="email" type="email" autocomplete="email" required
+					class="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500  rounded-t-md focus:outline-none focus:ring-gray-400 focus:border-gray-400 focus:z-10 text-md"
+					placeholder="Email">
 			</div>
-			<form class="mt-8 space-y-6" action="/login" method="POST">
-				<div class="rounded-md shadow-sm -space-y-px">
-					<div>
-						<label for="email" class="sr-only">Email</label>
-						<input id="email" name="email" type="email" autocomplete="email" required class="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500  rounded-t-md focus:outline-none focus:ring-gray-400 focus:border-gray-400 focus:z-10 text-md" placeholder="Email">
-					</div>
-					<div>
-						<label for="password" class="sr-only">Password</label>
-						<input id="password" name="password" type="password"required class="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500  rounded-b-md focus:outline-none focus:ring-gray-400 focus:border-gray-400 focus:z-10 text-md" placeholder="Password">
-					</div>
-				</div>
-				<div>
-					<button type="submit" class="text-white group relative w-full flex justify-center py-3 px-4 border border-transparent text-md font-medium rounded-md  bg-zinc-800 hover:bg-zinc-700 focus:outline-none focus:ring-2 focus:ring-gray-400">
-						<span class="absolute left-0 inset-y-0 flex items-center pl-3">
-							<svg class="h-5 w-5 " xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" aria-hidden="true">
-							<path fill-rule="evenodd" d="M5 9V7a5 5 0 0110 0v2a2 2 0 012 2v5a2 2 0 01-2 2H5a2 2 0 01-2-2v-5a2 2 0 012-2zm8-2v2H7V7a3 3 0 016 0z" clip-rule="evenodd" />
-							</svg>
-						</span>
-						Login
-					</button>
-				</div>
-			</form>
+			<div>
+				<label for="password" class="sr-only">Password</label>
+				<input id="password" name="password" type="password" required
+					class="appearance-none rounded-none relative block w-full px-3 py-2 border border-gray-300 placeholder-gray-500  rounded-b-md focus:outline-none focus:ring-gray-400 focus:border-gray-400 focus:z-10 text-md"
+					placeholder="Password">
+			</div>
 		</div>
-{{template "footer"}}
+		<div>
+			<button type="submit"
+				class="text-white group relative w-full flex justify-center py-3 px-4 border border-transparent text-md font-medium rounded-md  bg-zinc-800 hover:bg-zinc-700 focus:outline-none focus:ring-2 focus:ring-gray-400">
+				<span class="absolute left-0 inset-y-0 flex items-center pl-3">
+					<svg class="h-5 w-5 " xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor"
+						aria-hidden="true">
+						<path fill-rule="evenodd"
+							d="M5 9V7a5 5 0 0110 0v2a2 2 0 012 2v5a2 2 0 01-2 2H5a2 2 0 01-2-2v-5a2 2 0 012-2zm8-2v2H7V7a3 3 0 016 0z"
+							clip-rule="evenodd" />
+					</svg>
+				</span>
+				Login
+			</button>
+		</div>
+	</form>
+</div>
+{{end}}
