+++ app/views/pages/post.gohtml
@@ -2,7 +2,7 @@
 <!-- Post Section -->
 <script src="/static/prism/prism.js"></script>
 <section id="postWrap" class="content mx-auto">
-	<article class="serif flex flex-col max-w-4xl mx-auto text-zinc-800 p-2 sm:p-4 rounded-2xl bg-white shadow">
+	<article class="serif flex flex-col mx-auto text-zinc-800 p-2 sm:p-4 rounded-2xl bg-white shadow">
 		<div class="flex gap-2 mb-4 mt-2 sm:mt-0 px-2">
 			{{if .Article.Tags}}
 			{{ $len := len .Article.Tags }}
