+++ app/views/pages/register.gohtml
@@ -1,5 +1,5 @@
-{{template "header"}}
-			<div>
+{{ define "register"}}
+<div>
 				<h1 class="mt-6 text-center text-3xl font-extrabold text-gray-900">
 					Register
 				</h1>
@@ -32,4 +32,4 @@
 					</button>
 				</div>
 			</form>
-{{template "footer"}}
+{{ end }}
